updates for Location component tests
