+++ package.json
@@ -48,6 +48,7 @@
     "babel-register": "^6.26.0",
     "jest": "^26.2.2",
     "jest-axe": "^3.5.0",
-    "react-axe": "^3.5.3"
+    "react-axe": "^3.5.3",
+    "redux-mock-store": "^1.5.4"
   }
 }
